Only removing tmp.data if it really exists.
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@6565 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ scripts/results/quantilePlot.sh
@@ -46,7 +46,12 @@ CMD="plot ";
 k=0;
 for i in `ls $DIR`; do
   echo "$k: $i";
-  rm tmp.data
+
+  if [ -f "tmp.data" ]
+  then
+      rm tmp.data
+  fi
+
   CMD="$CMD 'tab_$i.data' using 1:2 with linespoints linecolor rgb \"${COLORS[$k]}\" pointinterval -10 linewidth 3.5 title '$i',";
   for j in `ls $DIR/$i/*.csv`; do
